+++ tests/testthat/test_make_eml.R
@@ -112,6 +112,26 @@ testthat::test_that('Expect argument values in EML', {
       x1$data.table.url[i])
   }
   
+  # data.table.url - A URL is not required for each dataTable
+  
+  x1 <- x
+  x1$data.table.url[1] <- ""
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$dataTable[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$dataTable[[2]]$physical$distribution$online$url,
+    x1$data.table.url[2])
+  
+  x1 <- x
+  x1$data.table.url[1] <- NA_character_
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$dataTable[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$dataTable[[2]]$physical$distribution$online$url,
+    x1$data.table.url[2])
+  
   # other.entity
   
   x1 <- x
@@ -120,6 +140,15 @@ testthat::test_that('Expect argument values in EML', {
   for (i in 1:length(r$dataset$otherEntity)) {
     expect_true(
       r$dataset$otherEntity[[i]]$physical$objectName == names(x1$x$other.entity[i]))
+  }
+  
+  # other.entity - MIME Type
+  
+  x1 <- x
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  for (i in 1:length(r$dataset$otherEntity)) {
+    expect_true(
+      r$dataset$otherEntity[[i]]$physical$dataFormat$externallyDefinedFormat$formatName != "Unknown")
   }
   
   # other.entity.name - defaults to other.entity
@@ -141,6 +170,26 @@ testthat::test_that('Expect argument values in EML', {
       r$dataset$otherEntity[[i]]$physical$distribution$online$url,
       x1$other.entity.url[i])
   }
+  
+  # other.entity.url - A URL is not required for each otherEntity
+  
+  x1 <- x
+  x1$other.entity.url[1] <- ""
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$otherEntity[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$otherEntity[[2]]$physical$distribution$online$url,
+    x1$other.entity.url[2])
+  
+  x1 <- x
+  x1$other.entity.url[1] <- NA_character_
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$otherEntity[[1]]$physical$distribution$online$url))
+  expect_equal(
+    r$dataset$otherEntity[[2]]$physical$distribution$online$url,
+    x1$other.entity.url[2])
   
   # provenance - Get provenance metadata for EDI data packages and place under
   # /eml/dataset/methods/methodStep/dataSource
@@ -181,6 +230,15 @@ testthat::test_that('Expect template values in EML', {
   r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
   expect_true(
     length(r$dataset$coverage$geographicCoverage) == nrow(x1$x$template$geographic_coverage.txt$content))
+  
+  # personnel.txt - Principal Investigator is absent
+  
+  x1 <- x
+  use_i <- which(x1$x$template$personnel.txt$content$role == "PI")
+  x1$x$template$personnel.txt$content <- x1$x$template$personnel.txt$content[-use_i, ]
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    is.null(r$dataset$project))
   
   # personnel.txt - only projectTitle is present
   
@@ -282,6 +340,26 @@ testthat::test_that('Expect template values in EML', {
     r$dataset$project$relatedProject[[4]]$funding,
     "No funding to report")
   
+  # personnel.txt - publisher
+  
+  x1 <- x
+  use_i <- min(which(x1$x$template$personnel.txt$content$role == "creator"))
+  x1$x$template$personnel.txt$content$role[use_i] <- "publisher"
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    !is.null(r$dataset$publisher))
+  expect_true(
+    length(r$dataset$publisher[[1]]) > 1)
+  
+  x1 <- x
+  use_i <- which(x1$x$template$personnel.txt$content$role == "creator")[1:3]
+  x1$x$template$personnel.txt$content$role[use_i] <- "publisher"
+  r <- do.call(make_eml, x1[names(x1) %in% names(formals(make_eml))])
+  expect_true(
+    !is.null(r$dataset$publisher))
+  expect_true(
+    length(r$dataset$publisher[[1]]) > 1)
+
   # taxonomic_coverage.txt
   
   x1 <- x